add start date and race count properties to leagues

+++ src/main/java/com/sss/garage/model/event/EventRepository.java
@@ -8,7 +8,7 @@ import java.util.Date;
 
 @Repository
 public interface EventRepository extends JpaRepository<Event, Long> {
-    Long countByLeague(League league);
+    Integer countByLeague(League league);
 
     Event findFirstByLeagueOrderByStartDateAsc(League league);
 }
